backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
better-exceptions==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0